requirements.txt
Flask
line-bot-sdk
beautifulsoup4
requests
selenium
webdriver-manager
python-dotenv
